% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_simulated_annealing.R
\name{adaptive.simulated.annealing}
\alias{adaptive.simulated.annealing}
\title{Adaptive simulated annealing}
\usage{
adaptive.simulated.annealing(poset, obs, lambda, eps, times = NULL,
  lambda.s = 1, weights = NULL, L, sampling = c("forward", "add-remove",
  "rejection"), version = 3L, max.iter = 100L, update.step.size = 20L,
  tol = 0.001, max.lambda.val = 1e+06, T0 = 50, adap.rate = 0.3,
  acceptance.rate = NULL, step.size = NULL, max.iter.asa = 10000L,
  adaptive = TRUE, outdir = NULL, thrds = 1L, verbose = FALSE,
  seed = NULL)
}
\arguments{
\item{poset}{a matrix containing the cover relations}

\item{obs}{a matrix of observations, where each row correponds to a
vector indicating whether an event has been observed (\code{1}) or not
(\code{0})}

\item{lambda}{a vector of the rate parameters}

\item{eps}{error rate}

\item{times}{an optional vector of sampling times per observation}

\item{lambda.s}{rate of the sampling process. Defaults to \code{1.0}}

\item{weights}{a vector containing observation weights}

\item{L}{number of samples to be drawn from the proposal in the E-step}

\item{sampling}{sampling scheme. OPTIONS: \code{"forward"} - generate
occurrence times according to current rate parameters, and, from them,
generate the genotypes; \code{"add-remove"} - generate genotypes from
observed genotypes using a two-steps proposal. First, make genotypes
compatible with the poset by either adding or removing mutations. Second,
perturb this version by adding or removing a mutation while yielding a
compatible observation; \code{"rejection"} - generate a pool of compatible
genotypes according to current rate parameters and sample \code{K}
observations proportional to the Hamming distance.}

\item{version}{an optional arguement indicating which version of the
\code{"add-remove"} sampling scheme to use. Defaults to \code{3}}

\item{max.iter}{the maximum number of EM iterations. Defaults to \code{100}
iterations}

\item{update.step.size}{number of EM steps after which the number of
samples, \code{L}, is doubled. \code{L} is increased, if the difference in
the parameter estimates between such consecutive batches is greater than the
tolerance level, \code{tol}}

\item{tol}{convergence tolerance for the error rate and the rate parameters.
The EM runs until the difference between the average estimates in the last
two batches is smaller than tol, or until \code{max.iter} is reached.}

\item{max.lambda.val}{an optional upper bound on the value of the rate
parameters. Defaults to \code{1e6}}

\item{T0}{inital value of the temperature. Defaults to \code{50}}

\item{adap.rate}{an optional argument specifying the constant adaptation
rate. Defaults to \code{0.3}}

\item{acceptance.rate}{an optional argument specifying a desirable
acceptance rate. Defaults to 1/\code{p}, where \code{p} is the number of
events or mutations}

\item{step.size}{an optional argument indicating the number or iterations
after which the temperature should be updated. Defaults to \code{50}}

\item{max.iter.asa}{the maximum number of iterations. Defaults to
\code{10000} iterations}

\item{adaptive}{a boolean variable indicating whether to use an adaptive
annealing schedule}

\item{outdir}{an optional argument indicating the path to the output
directory}

\item{thrds}{number of threads for parallel execution}

\item{verbose}{an optional argument indicating whether to output logging
information}

\item{seed}{seed for reproducibility}
}
\description{
structure learning using adaptive simulated annealing
}
