% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{complete.loglikelihood}
\alias{complete.loglikelihood}
\title{Complete-data Log-Likelihood}
\usage{
complete.loglikelihood(lambda, eps, Tdiff, dist, weights = NULL)
}
\arguments{
\item{lambda}{a vector of the rate parameters}

\item{eps}{error rate, \eqn{\epsilon}}

\item{Tdiff}{a matrix of expected time differences}

\item{dist}{a vector of expected Hamming distances}

\item{weights}{an optional vector containing observation weights}
}
\description{
compute the complete-data log-likelihood or, equivalently, the
hidden log-likelihood
}
