% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/param_est.r
\name{estimate_lambda}
\alias{estimate_lambda}
\title{estimate_lambda}
\usage{
estimate_lambda(obs_events, dtimes, poset_, likelihood_fn = likelihood,
  init_lambda = NA, init_eps = 0.3, grad = NULL, optim_method = "minqa",
  control = list(), with_eps = FALSE, ...)
}
\description{
estimate_lambda
}

