% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/param_est.r
\name{fit_params}
\alias{fit_params}
\title{fit_params}
\usage{
fit_params(poset, obs_events, times, D = NA, ilambda = NA,
  use_grad = FALSE, optim_method = "minqa", control = list(),
  err_model = c("no-error", "hamming"), eps = NA, ...)
}
\description{
fit_params
}

