% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/poset_learning_pc.R
\name{fit_weibull}
\alias{fit_weibull}
\title{fit_weibull
B * N/4 is the maximum allwed threshold in the last iteration. The error threshold in iteration "iter" is defined
as max( 0.2, B * (N/4) * iter/max_iteration). The B * N/4 still gives a OK result. B * (N/4) is an approximation for
B * fa * fb/ (fa+fb) where fa and fb are the frequncy of events a and b. The derivation of this formula is straightforward.
0.1 is a reasonable value for this parameter.}
\usage{
fit_weibull(E, T, org_init_par = NULL, B = 0.1, verbose = F)
}
\description{
fit_weibull
B * N/4 is the maximum allwed threshold in the last iteration. The error threshold in iteration "iter" is defined
as max( 0.2, B * (N/4) * iter/max_iteration). The B * N/4 still gives a OK result. B * (N/4) is an approximation for
B * fa * fb/ (fa+fb) where fa and fb are the frequncy of events a and b. The derivation of this formula is straightforward.
0.1 is a reasonable value for this parameter.
}

