% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_genotypes.R
\name{generate.genotypes}
\alias{generate.genotypes}
\title{Generate genotypes}
\usage{
generate.genotypes(
  mutation.times,
  poset,
  sampling.time = NULL,
  lambda.s = 1,
  seed = NULL
)
}
\arguments{
\item{mutation.times}{a matrix of dimension (\code{n} x \code{p}), where
\code{n} is the number of genotypes to be generated and \code{p} is the
number of events (or mutations)}

\item{poset}{a matrix containing the cover relations}

\item{sampling.time}{an optional argument specifying the sampling time}

\item{lambda.s}{rate of the sampling process. Defaults to \code{1.0}}

\item{seed}{seed for reproducibility}
}
\description{
Generate observations from mutations times
}
