% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcem_hcbn.R
\name{importance.weight}
\alias{importance.weight}
\title{Importance sampling}
\usage{
importance.weight(genotype, L, poset, lambda, eps, time = NULL,
  sampling = c("forward", "add-remove", "rejection"), version = NULL,
  lambda.s = 1, thrds = 1L, seed = NULL)
}
\arguments{
\item{genotype}{either a binary vector or a matrix containing observations
or genotypes. Each row of the matrix corresponds to a genotype vector whose
entries indicate whether an event has been observed (\code{1}) or not
(\code{0})}

\item{L}{number of samples to be drawn from the proposal}

\item{poset}{a matrix containing the cover relations}

\item{lambda}{a vector of the rate parameters}

\item{eps}{error rate}

\item{time}{optional argument specifying the sampling time}

\item{sampling}{sampling scheme. OPTIONS: \code{"forward"} - generate
occurrence times according to current rate parameters, and, from them,
generate the genotypes; \code{"add-remove"} - generate genotypes from
observed genotypes using a two-steps proposal. First, make genotypes
compatible with the poset by either adding or removing mutations. Second,
perturb this version by adding or removing a mutation while yielding a
compatible observation; \code{"rejection"} - generate a pool of compatible
genotypes according to current rate parameters and sample \code{K}
observations proportional to the Hamming distance}

\item{version}{an integer indicating which version of the
\code{"add-remove"} sampling scheme to use}

\item{lambda.s}{rate of the sampling process. Defaults to \code{1.0}}

\item{thrds}{number of threads for parallel execution. This option is used
if \code{genotype} corresponds to a matrix of genotypes}

\item{seed}{seed for reproducibility}
}
\description{
compute the sufficient statistics in expectation using
importance sampling
}
