% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbn_sampler.r
\name{make_empty_poset}
\alias{make_empty_poset}
\title{Making an empty poset}
\usage{
make_empty_poset(p)
}
\arguments{
\item{p}{the number of events}
}
\value{
the empty poset of size \code{p}
}
\description{
This function creates an empty poset with p nodes and no edges.
}
\examples{
make_empty_poset(4)
}
