% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccbn-package.R
\docType{package}
\name{mccbn-package}
\alias{mccbn-package}
\title{mccbn}
\description{
\emph{mccbn} is a package that performs structure learning and parameter
estimation for large-scale continuous-time conjunctive Bayesian networks.
Two error models are implemented: a mixture and a latent model. Furthermore,
if available, sequencing times can be incorporate into the method.
}
\author{
Hesam Montazeri

Susana Posada Céspedes | \email{Susana Posada Céspedes}
}
