% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{obs.loglikelihood}
\alias{obs.loglikelihood}
\title{Observed Log-Likelihood}
\usage{
obs.loglikelihood(obs, poset, lambda, eps, times = NULL, L,
  sampling = c("forward", "add-remove", "pool", "backward"),
  neighborhood.dist = 1L, lambda.s = 1, thrds = 1L, seed = NULL)
}
\arguments{
\item{obs}{a matrix containing observations or genotypes, where each row
correponds to a genotype vector whose entries indicate whether an event has
been observed (\code{1}) or not (\code{0})}

\item{poset}{a matrix containing the cover relations}

\item{lambda}{a vector of the rate parameters}

\item{eps}{error rate}

\item{times}{an optional vector of sampling times per observation}

\item{L}{number of samples to be drawn from the proposal}

\item{sampling}{type of sampling scheme. OPTIONS: \code{"forward"},
\code{"add-remove"}, \code{"pool"} or \code{"backward"}}

\item{neighborhood.dist}{an integer value indicating the Hamming distance
between the observation and the samples generated by \code{"backward"}
sampling. This option is used if \code{sampling} is set to \code{"backward"}.
Defaults to \code{1}}

\item{lambda.s}{rate of the sampling process. Defaults to \code{1.0}}

\item{thrds}{number of threads for parallel execution}

\item{seed}{seed for reproducibility}
}
\description{
compute the observed log-likelihood
}
