% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_genotypes.R
\name{sample.genotypes}
\alias{sample.genotypes}
\title{Sample genotypes}
\usage{
sample.genotypes(
  N,
  poset,
  lambda,
  sampling.times = NULL,
  lambda.s = 1,
  seed = NULL
)
}
\arguments{
\item{N}{number of samples}

\item{poset}{a matrix containing the cover relations}

\item{lambda}{a vector of the rate parameters}

\item{sampling.times}{an optional vector of sampling times per observation}

\item{lambda.s}{rate of the sampling process. Defaults to \code{1.0}}

\item{seed}{seed for reproducibility}
}
\description{
Generate observations from a given poset and given rates
}
