% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_genotypes.R
\name{sample.times}
\alias{sample.times}
\title{Sample mutation times}
\usage{
sample.times(N, poset, lambda, seed = NULL)
}
\arguments{
\item{N}{number of samples}

\item{poset}{a matrix containing the cover relations}

\item{lambda}{a vector of the rate parameters}

\item{seed}{seed for reproducibility}
}
\description{
Generate observation times from a given poset and given rates
}
