% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcem_hcbn.R
\name{importance.weight}
\alias{importance.weight}
\title{Importance sampling}
\usage{
importance.weight(genotype, L, poset, lambda, eps, time = NULL,
  sampling = c("forward", "add-remove", "pool", "backward"),
  weight.remove = numeric(0), dist.pool = integer(0),
  Tdiff.pool = matrix(0), neighborhood.dist = 1L, lambda.s = 1,
  thrds = 1L, seed = NULL)
}
\arguments{
\item{genotype}{either a binary vector or a matrix containing observations
or genotypes. Each row of the matrix corresponds to a genotype vector whose
entries indicate whether an event has been observed (\code{1}) or not
(\code{0})}

\item{L}{number of samples to be drawn from the proposal}

\item{poset}{a matrix containing the cover relations}

\item{lambda}{a vector of the rate parameters}

\item{eps}{error rate}

\item{time}{optional argument specifying the sampling time}

\item{sampling}{sampling scheme. OPTIONS: \code{"forward"} - generate
occurrence times according to current rate parameters, and, from them,
generate the genotypes; \code{"add-remove"} - generate genotypes from
observed genotypes using a two-steps proposal. First, make genotypes
compatible with the poset by either adding or removing mutations. Second,
perturb this version by adding or removing a mutation while yielding a
compatible observation; \code{"pool"} - generate a pool of compatible
genotypes according to current rate parameters and sample \code{K}
observations proportional to the Hamming distance; \code{"backward"} -
generate genotypes with Hamming distance \code{k}}

\item{weight.remove}{a numeric vector of length \code{p} containing the
weights for choosing events to be removed. This option is used if
\code{sampling} is set to \code{"add-remove"}}

\item{dist.pool}{Hamming distance between \code{genotype} and the genotype
pool. This option is used if \code{sampling} is set to \code{"pool"}
and \code{genotype} corresponds to a vector containing a single genotype}

\item{Tdiff.pool}{Expected time differences for the genotype pool. This
option is used if \code{sampling} is set to \code{"pool"} and
\code{genotype} corresponds to a vector containing a single genotype}

\item{neighborhood.dist}{an integer value indicating the Hamming distance
between the observation and the samples generated by \code{"backward"}
sampling. This option is used if \code{sampling} is set to \code{"backward"}.
Defaults to \code{1}}

\item{lambda.s}{rate of the sampling process. Defaults to \code{1.0}}

\item{thrds}{number of threads for parallel execution. This option is used
if \code{genotype} corresponds to a matrix of genotypes}

\item{seed}{seed for reproducibility}
}
\description{
compute the sufficient statistics in expectation using
importance sampling
}
