% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/likelihood.r
\name{loglike_mixture_model}
\alias{loglike_mixture_model}
\title{loglike_mixture_model}
\usage{
loglike_mixture_model(poset, lambda, obs_events, times, weights,
  control = list(ll_method = "importance"))
}
\description{
loglike_mixture_model
}

