% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcem_hcbn.R
\name{MCEM.hcbn}
\alias{MCEM.hcbn}
\title{Monte Carlo Expectation Maximization}
\usage{
MCEM.hcbn(lambda, poset, obs, lambda.s = 1, L, eps = NULL,
  sampling = c("forward", "add-remove", "rejection"), times = NULL,
  weights = NULL, version = 3L, max.iter = 100L, update.step.size = 20L,
  tol = 0.001, max.lambda = 1e+06, thrds = 1L, verbose = FALSE,
  seed = NULL)
}
\arguments{
\item{lambda}{a vector containing initial values for the rate parameters}

\item{poset}{a matrix containing the cover relations}

\item{obs}{a matrix containing observations or genotypes, where each row
corresponds to a genotype vector whose entries indicate whether an event has
been observed (\code{1}) or not (\code{0})}

\item{lambda.s}{rate of the sampling process. Defaults to \code{1.0}}

\item{L}{number of samples to be drawn from the proposal in the E-step}

\item{eps}{an optional initial value of the error rate parameter}

\item{sampling}{sampling scheme. OPTIONS: \code{"forward"} - generate
occurrence times according to current rate parameters, and, from them,
generate the genotypes; \code{"add-remove"} - generate genotypes from
observed genotypes using a two-steps proposal. First, make genotypes
compatible with the poset by either adding or removing mutations. Second,
perturb this version by adding or removing a mutation while yielding a
compatible observation; \code{"rejection"} - generate a pool of compatible
genotypes according to current rate parameters and sample \code{K}
observations proportional to the Hamming distance}

\item{times}{an optional vector containing times at which genotypes were
observed}

\item{weights}{an optional vector containing observation weights}

\item{version}{an optional argument indicating which version of the
\code{"add-remove"} sampling scheme to use. Defaults to \code{3}}

\item{max.iter}{the maximum number of EM iterations. Defaults to \code{100}
iterations}

\item{update.step.size}{number of EM steps after which the number of
samples, \code{L}, is doubled. \code{L} is increased, if the difference in
the parameter estimates between such consecutive batches is greater than the
tolerance level, \code{tol}}

\item{tol}{convergence tolerance for the error rate and the rate parameters.
The EM runs until the difference between the average estimates in the last
two batches is smaller than tol, or until \code{max.iter} is reached.}

\item{max.lambda}{an optional upper bound on the value of the rate
parameters. Defaults to \code{1e6}}

\item{thrds}{number of threads for parallel execution}

\item{verbose}{an optional argument indicating whether to output logging
information}

\item{seed}{seed for reproducibility}
}
\description{
parameter estimation for the hidden conjunctive Bayesian network
model (H-CBN) via importance sampling
}
