% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cbn_sampler.r
\name{sample_genotypes}
\alias{sample_genotypes}
\title{sample_timed_genotypes_with_eps}
\usage{
sample_genotypes(n, poset, sampling_param, lambdas, sampling_fn, eps = 0)
}
\description{
sample_timed_genotypes_with_eps
}

